these stinking int types were messing with me all along!

+++ aoc6/aoc62/FishAutomata.h
@@ -4,9 +4,9 @@
 class FishAutomata
 {
 public:
-	std::vector<unsigned long> initialState;
-	std::vector<unsigned long> getState();
-	void alterState(std::vector<unsigned long>& v);
+	std::vector<unsigned long long> initialState;
+	std::vector<unsigned long long> getState();
+	void alterState(std::vector<unsigned long long>& v);
 	void setInitialState(std::vector<int>&);
 };
